requirements.txt
streamlit
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
pypdf
tiktoken
python-dotenv